# docs/openapi.yaml
openapi: 3.0.3
info:
  title: LumeScope API
  version: v1
  description: |
    LumeScope API provides access to Lumera blockchain action data, supernode metrics, 
    and version compatibility information. This API enables monitoring and analysis of 
    the Lumera network's distributed storage operations.
servers:
  - url: /
    description: Current server
paths:
...
  /v1/actions:
    get:
      summary: List actions
      description: Retrieve a paginated list of blockchain actions with optional filtering
      operationId: listActions
      tags:
        - Actions
      parameters:
        - name: type
          in: query
          description: Filter by action type
          schema:
            type: string
            enum:
              - ACTION_TYPE_UNSPECIFIED
              - ACTION_TYPE_CASCADE
              - ACTION_TYPE_SENSE
        - name: creator
          in: query
          description: Filter by creator address
          schema:
            type: string
        - name: state
          in: query
          description: Filter by action state
          schema:
            type: string
            enum:
              - ACTION_STATE_UNSPECIFIED
              - ACTION_STATE_PENDING
              - ACTION_STATE_DONE
              - ACTION_STATE_APPROVED
              - ACTION_STATE_EXPIRED
        - name: from
          in: query
          description: Filter actions from block height
          schema:
            type: integer
            format: int64
        - name: to
          in: query
          description: Filter actions to block height
          schema:
            type: integer
            format: int64
        - name: limit
          in: query
          description: Maximum number of items to return (1-200)
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
        - name: cursor
          in: query
          description: Pagination cursor for next page
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionsListResponse'
  /v1/actions/{id}:
    get:
      summary: Get action details
      description: Retrieve detailed information about a specific action by ID
      operationId: getAction
      tags:
        - Actions
      parameters:
        - name: id
          in: path
          required: true
          description: Action ID
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionDetail'
        '404':
          description: Action not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/actions/stats:
    get:
      summary: Get global action statistics
      description: |
        Retrieve aggregated action statistics for all actions (global).
        Returns total count and a breakdown by state (PENDING, DONE, APPROVED, EXPIRED).
      operationId: getActionStats
      tags:
        - Actions
      parameters:
        - name: type
          in: query
          description: Filter by action type
          schema:
            type: string
            enum:
              - ACTION_TYPE_UNSPECIFIED
              - ACTION_TYPE_CASCADE
              - ACTION_TYPE_SENSE
      responses:
        '200':
          description: Successful response with action statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionStatsResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /v1/supernodes/unavailable:
    get:
      summary: List unavailable supernodes
      description: Retrieve supernodes where status API is unavailable (isStatusApiAvailable=false), with optional state filtering
      operationId: listUnavailableSupernodes
      tags:
        - Supernodes
      parameters:
        - name: currentState
          in: query
          description: Filter by current state of supernodes
          schema:
            type: string
            enum:
              - running
              - stopped
              - any
            default: running
      responses:
        '200':
          description: Successful response with list of unavailable supernodes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SupernodeDetail'
        '400':
          description: Invalid query parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /v1/supernodes/stats:
    get:
      summary: Get aggregated supernode hardware statistics
      description: |
        Retrieve aggregated hardware statistics for fully available supernodes.
        A supernode is considered "fully available" when all three ports are open:
        - Status API (port 8002)
        - Port1 (from ipAddress)
        - P2P port (4445)
      operationId: getSupernodeStats
      tags:
        - Supernodes
      responses:
        '200':
          description: Successful response with aggregated hardware stats
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupernodeStatsResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/supernodes/action-stats:
    get:
      summary: Get action statistics for a supernode
      description: |
        Retrieve aggregated action statistics for a specific supernode address.
        Returns counts grouped by action state for all actions where the supernode
        is listed in the superNodes array.
      operationId: getSupernodeActionStats
      tags:
        - Supernodes
      parameters:
        - name: address
          in: query
          required: true
          description: The supernode address to filter actions by
          schema:
            type: string
            example: lumera1563uuzljqvpanh79w2tvymqsly73v9nwygda73
        - name: type
          in: query
          description: Filter by action type
          schema:
            type: string
            enum:
              - ACTION_TYPE_UNSPECIFIED
              - ACTION_TYPE_CASCADE
              - ACTION_TYPE_SENSE
      responses:
        '200':
          description: Successful response with action statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupernodeActionStatsResponse'
        '400':
          description: Invalid or missing address parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    ActionStatsResponse:
      type: object
      description: Aggregated action statistics for all actions (global)
      properties:
        total:
          type: integer
          description: Total number of actions
          example: 100
        states:
          type: object
          description: Count of actions grouped by state
          additionalProperties:
            type: integer
          example:
            ACTION_STATE_PENDING: 10
            ACTION_STATE_DONE: 80
            ACTION_STATE_APPROVED: 5
            ACTION_STATE_EXPIRED: 5
        schema_version:
          type: string
          description: API schema version
          example: v1.0
      required:
        - total
        - states
        - schema_version
    SupernodeActionStatsResponse:
      type: object
      description: Aggregated action statistics for a specific supernode
      properties:
        total:
          type: integer
          description: Total number of actions associated with the supernode
          example: 10
        states:
          type: object
          description: Count of actions grouped by state
          additionalProperties:
            type: integer
          example:
            ACTION_STATE_PENDING: 2
            ACTION_STATE_DONE: 8
        supernode_address:
          type: string
          description: The supernode address used for filtering
          example: lumera1563uuzljqvpanh79w2tvymqsly73v9nwygda73
        schema_version:
          type: string
          description: API schema version
          example: v1.0
      required:
        - total
        - states
        - supernode_address
        - schema_version
    SupernodeStatsResponse:
      type: object
      description: Aggregated hardware statistics for fully available supernodes
      properties:
        total_cpu_cores:
          type: integer
          format: int64
          description: Sum of CPU cores across all fully available supernodes
          example: 256
        total_memory_gb:
          type: number
          format: double
          description: Sum of total memory (GB) across all fully available supernodes
          example: 1024.5
        total_storage_bytes:
          type: integer
          format: int64
          description: Sum of total storage (bytes) across all fully available supernodes
          example: 10995116277760
        used_storage_bytes:
          type: integer
          format: int64
          description: Sum of used storage (bytes) across all fully available supernodes
          example: 5497558138880
        available_storage_bytes:
          type: integer
          format: int64
          description: Calculated available storage (total - used) in bytes
          example: 5497558138880
        storage_used_percent:
          type: number
          format: double
          description: Percentage of storage used (used / total * 100)
          example: 50.0
        storage_available_percent:
          type: number
          format: double
          description: Percentage of storage available (available / total * 100)
          example: 50.0
        available_supernodes:
          type: integer
          format: int64
          description: Number of fully available supernodes included in the statistics
          example: 42
        schema_version:
          type: string
          description: API schema version
          example: v1.0
    ActionsListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ActionItem'
        next_cursor:
          type: string
          description: Cursor for fetching the next page
        schema_version:
          type: string
          example: v1.0
    ActionItem:
      type: object
      properties:
        id:
          type: string
          description: Unique action identifier
          example: act_1
        type:
          type: string
          description: Action type (e.g., cascade, transfer)
          example: cascade
        creator:
          type: string
          description: Creator address
          example: lumera1example...
        state:
          type: string
          description: Current action state
          example: committed
        block_height:
          type: integer
          format: int64
          description: Block height when action was created
          example: 12345
        decoded:
          type: object
          additionalProperties: true
          description: Decoded action data
        raw:
          type: string
          description: Base64 encoded raw bytes for unknown types
        superNodes:
          type: array
          description: List of SuperNode operator addresses associated with the action
          items:
            type: string
            example: lumera1563uuzljqvpanh79w2tvymqsly73v9nwygda73
    ActionDetail:
      type: object
      properties:
        id:
          type: string
          description: Unique action identifier
        type:
          type: string
          description: Action type
          example: cascade
        creator:
          type: string
          description: Creator address
        state:
          type: string
          description: Current action state
          example: committed
        block_height:
          type: integer
          format: int64
          description: Block height when action was created
          example: 12345
        price:
          $ref: '#/components/schemas/Price'
        timestamp:
          type: string
          format: date-time
          description: Action timestamp
        decoded:
          type: object
          additionalProperties: true
          description: Decoded action data
        raw:
          type: string
          description: Base64 encoded raw bytes for unknown types
        super_nodes:
          type: object
          additionalProperties: true
          description: SuperNode data associated with the action
        schema_version:
          type: string
          example: v1.0
    Price:
      type: object
      description: Price information for an action
      properties:
        denom:
          type: string
          description: Price denomination
          example: ulume
        amount:
          type: string
          description: Price amount as string
          example: "1000000"
...
    Error:
      type: object
      description: Error response
      properties:
        error:
          type: string
          description: Error message
          example: internal_error
      required:
        - error
    SupernodeDetail:
      type: object
      properties:
        supernodeAccount:
          type: string
          description: Supernode account address
        validatorAddress:
          type: string
          description: Validator address
        validatorMoniker:
          type: string
          description: Validator moniker/name
        currentState:
          type: string
          description: Current state of the supernode
        ipAddress:
          type: string
          description: IP address of the supernode
        p2pPort:
          type: integer
          description: P2P port number
        protocolVersion:
          type: string
          description: Protocol version
        isStatusApiAvailable:
          type: boolean
          description: Whether the status API is available
...